creating login and private routes

+++ src/actions/peliculas.js
@@ -1,5 +1,7 @@
-import { FetchGetData } from "../helpers/fetch";
-
+import {
+    FetchGetData
+} from "../helpers/fetch";
+import Swal from 'sweetalert2'
 const {
     types
 } = require("../types/types");
@@ -11,7 +13,7 @@ export const StartGetAllMovies = () => {
             const resp = await fetch("https://movieweb-react.herokuapp.com/peliculas")
             const data = await resp.json()
             // console.log(data.peliculas)
-            const peliculas=data.peliculas
+            const peliculas = data.peliculas
             dispatch(getAllMovies(peliculas))
         } catch (error) {
             console.log(error)
@@ -25,122 +27,121 @@ const getAllMovies = (peliculas) => ({
     payload: peliculas
 })
 
-export const StartGetProximas=()=>{
+export const StartGetProximas = () => {
 
-    return async(dispatch)=>{
-        const data=await FetchGetData()
-        const proximas=data.filter(pelicula=>pelicula.seccion==="proximas")
+    return async (dispatch) => {
+        const data = await FetchGetData()
+        const proximas = data.filter(pelicula => pelicula.seccion === "proximas")
         dispatch(getProximas(proximas))
     }
 }
 
-const getProximas =(peliculas)=>({
-    type:types.getProximas,
-    payload:peliculas
+const getProximas = (peliculas) => ({
+    type: types.getProximas,
+    payload: peliculas
 })
 
-export const StartGetAnimes=()=>{
+export const StartGetAnimes = () => {
 
-    return async(dispatch)=>{
-        const data=await FetchGetData()
-        const anime=data.filter(pelicula=>pelicula.seccion==="anime")
+    return async (dispatch) => {
+        const data = await FetchGetData()
+        const anime = data.filter(pelicula => pelicula.seccion === "anime")
         dispatch(getAnime(anime))
     }
 }
 
-const getAnime =(peliculas)=>({
-    type:types.getAnimes,
-    payload:peliculas
+const getAnime = (peliculas) => ({
+    type: types.getAnimes,
+    payload: peliculas
 })
 
-export const StartGetSeries=()=>{
+export const StartGetSeries = () => {
 
-    return async(dispatch)=>{
-        const data=await FetchGetData()
-        const series = data.filter(pelicula=>pelicula.seccion==="series")
+    return async (dispatch) => {
+        const data = await FetchGetData()
+        const series = data.filter(pelicula => pelicula.seccion === "series")
         dispatch(getSeries(series))
     }
 }
 
-const getSeries=(series)=>({
-    type:types.getSeries,
-    payload:series
+const getSeries = (series) => ({
+    type: types.getSeries,
+    payload: series
 })
 
-export const StartGetNuevas=()=>{
+export const StartGetNuevas = () => {
 
-    return async(dispatch)=>{
-        const data=await FetchGetData()
-        const peliculas= data.filter(pelicula=>pelicula.seccion==="nuevas")
+    return async (dispatch) => {
+        const data = await FetchGetData()
+        const peliculas = data.filter(pelicula => pelicula.seccion === "nuevas")
         dispatch(getNuevas(peliculas))
     }
 }
 
-const getNuevas=(peliculas)=>({
-    type:types.getNuevas,
-    payload:peliculas
+const getNuevas = (peliculas) => ({
+    type: types.getNuevas,
+    payload: peliculas
 })
 
-export const StartGetMas=()=>{
+export const StartGetMas = () => {
 
-    return async(dispatch)=>{
-        const data=await FetchGetData()
-        const peliculas= data.filter(pelicula=>pelicula.seccion==="mas")
+    return async (dispatch) => {
+        const data = await FetchGetData()
+        const peliculas = data.filter(pelicula => pelicula.seccion === "mas")
         dispatch(getMas(peliculas))
     }
 }
 
-const getMas=(peliculas)=>({
-    type:types.getSeries,
-    payload:peliculas
+const getMas = (peliculas) => ({
+    type: types.getSeries,
+    payload: peliculas
 })
 
-export const StartGetTrailers=()=>{
-    return async(dispatch)=>{
-        
-       try {
-        
-        const data=await FetchGetData()
-        const trailers = data.filter(trailers=>trailers.trailerId)
+export const StartGetTrailers = () => {
+    return async (dispatch) => {
 
-        // console.log(trailers)
-        dispatch(getTrailers(trailers))
+        try {
+
+            const data = await FetchGetData()
+            const trailers = data.filter(trailers => trailers.trailerId)
+
+            // console.log(trailers)
+            dispatch(getTrailers(trailers))
 
 
-       } catch (error) {
-           console.log(error)
-       }
+        } catch (error) {
+            console.log(error)
+        }
     }
 }
 
-const getTrailers =(trailers)=>({
-    type:types.getTrailers,
-    payload:trailers
+const getTrailers = (trailers) => ({
+    type: types.getTrailers,
+    payload: trailers
 })
 
-export const StartGetPeliculaById=(id)=>{
-    return async (dispatch)=>{
-        const resp= await fetch(`https://movieweb-react.herokuapp.com/peliculas/${id}`)
+export const StartGetPeliculaById = (id) => {
+    return async (dispatch) => {
+        const resp = await fetch(`https://movieweb-react.herokuapp.com/peliculas/${id}`)
         const data = await resp.json()
-        const pelicula=data.pelicula
-    
-  
+        const pelicula = data.pelicula
+
+
         dispatch(getPeliculaById(pelicula))
     }
 }
 
-const getPeliculaById =(pelicula)=>({
-    type:types.getOneMovie,
-    payload:pelicula
+const getPeliculaById = (pelicula) => ({
+    type: types.getOneMovie,
+    payload: pelicula
 })
 
-export const StartGetBannerMovie=()=>
-{
-    return async(dispatch)=>{
+export const StartGetBannerMovie = () => {
+    return async (dispatch) => {
         try {
             const data = await FetchGetData()
-            const pelicula = data.filter(pelicula=>pelicula.seccion==="banner")
-           
+            const pelicula = data.filter(pelicula => pelicula.seccion === "banner")
+
             dispatch(getBannerMovie(pelicula))
         } catch (error) {
             console.log(error)
@@ -148,52 +149,83 @@ export const StartGetBannerMovie=()=>
     }
 }
 
-const getBannerMovie=(bannerMovie)=>({
-    type:types.getBanner,
-    payload:bannerMovie
+const getBannerMovie = (bannerMovie) => ({
+    type: types.getBanner,
+    payload: bannerMovie
 })
 
-export const StartCreateMovie=(imagen,data)=>{
+export const StartCreateMovie = (imagen, data) => {
 
-    return async(dispatch)=>{
+    return async (dispatch) => {
 
         try {
-           
-            data.imagen=imagen
-            
 
-            const resp = await fetch('http://movieweb-react.herokuapp.com/peliculas/nueva',{
-                method:'POST',
-                headers:{
-                    'Content-type':'application/json'
+            data.imagen = imagen
+
+            const resp = await fetch('http://movieweb-react.herokuapp.com/peliculas/nueva', {
+                method: 'POST',
+                headers: {
+                    'Content-type': 'application/json'
                 },
-                body:JSON.stringify(data)
+                body: JSON.stringify(data)
             })
 
-            const body= await resp.json()
+            const body = await resp.json()
             dispatch(createMovie(data))
+            Swal.fire({
+                icon: 'success',
+                title: body.msg,
+                showConfirmButton: false,
+                timer: 1500
+            })
             console.log(body.msg)
-            
+
         } catch (error) {
             console.log(error)
         }
     }
 }
 
-const createMovie=()=>({ type:types.createMovie})
+const createMovie = () => ({
+    type: types.createMovie
+})
 
-export const StartDelete=(id)=>{
-        return async(dispatch)=>{
-           
-            const resp = await fetch(`http://movieweb-react.herokuapp.com/peliculas/delete/${id}`,{
-                method:"DELETE"
+export const StartDelete = (id) => {
+    return (dispatch) => {
+        try {
+
+            Swal.fire({
+                title: 'Esta Seguro?',
+                text: "Si es borrado esta accion no se puede revertir!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Si,Borrar!'
+            }).then(async (result) => {
+                if (result.isConfirmed) {
+                    const resp = await fetch(`http://movieweb-react.herokuapp.com/peliculas/delete/${id}`, {
+                        method: "DELETE"
+                    })
+                    const body = await resp.json()
+                    console.log(body)
+                    dispatch(deleteMovie)
+
+                    Swal.fire(
+                        'Borrado!',
+                        'Se ha borrado satisfactoriamente.',
+                        'success'
+                    )
+                }
             })
-            const body = await resp.json()
-            console.log(body)
-            dispatch(deleteMovie)
+
+        } catch (error) {
+            console.log(error)
         }
+
+    }
 }
 
-const deleteMovie=()=>({
-    type:types.deleteMovie
+const deleteMovie = () => ({
+    type: types.deleteMovie
 })
